change JavaFXApplicationSettings to ApplicationSettings
